regex search is really janky
we treat regexes as part of the same code path as regular searches and add a bunch of conditional checks everywhere. this is a huge pain to maintain and debug. let's move this to a different top level code path, and switch on isRegex as soon as the query is received